Create main.py
Hi, this python file creates a figure that shows the linear relation between the Number of Followers and time.

Code simply collects the related data thanks to Twitter API, then creates a dataFrame with two parameters: #Followers & #Time. I've commented on Seaborn visualization as an option as well. Each time the .py file runs, a new line of data will add to the excel file. I've also added a cronjob task scheduler to my local PC. If this code is installed on a virtual machine and runs within time intervals, the desired dataset would gather.